Moved ComputeSamples and related functions to vmc_sampling
